+++ semgrep/semgrep/util.py
@@ -19,6 +19,7 @@ from urllib.parse import urlparse
 import click
 
 from semgrep.constants import Colors
+from semgrep.constants import USER_LOG_FILE
 from semgrep.constants import YML_SUFFIXES
 from semgrep.constants import YML_TEST_SUFFIXES
 
@@ -55,28 +56,40 @@ def is_url(url: str) -> bool:
         return False
 
 
-# TODO: seems dead
 def set_flags(*, verbose: bool, debug: bool, quiet: bool, force_color: bool) -> None:
     """Set the relevant logging levels"""
     # Assumes only one of verbose, debug, quiet is True
-
     logger = logging.getLogger("semgrep")
-    logger.handlers = []
-    handler = logging.StreamHandler()
-    formatter = logging.Formatter("%(message)s")
-    handler.setFormatter(formatter)
+    logger.handlers = []  # Reset to no handlers
 
-    level = logging.INFO
+    stdout_level = logging.INFO
     if verbose:
-        level = logging.VERBOSE  # type: ignore[attr-defined]
+        stdout_level = logging.VERBOSE  # type: ignore[attr-defined]
     elif debug:
-        level = logging.DEBUG
+        stdout_level = logging.DEBUG
     elif quiet:
-        level = logging.CRITICAL
+        stdout_level = logging.CRITICAL
 
-    handler.setLevel(level)
-    logger.addHandler(handler)
-    logger.setLevel(level)
+    # Setup stdout logging
+    stdout_handler = logging.StreamHandler()
+    stdout_formatter = logging.Formatter("%(message)s")
+    stdout_handler.setFormatter(stdout_formatter)
+    stdout_handler.setLevel(stdout_level)
+    logger.addHandler(stdout_handler)
+
+    # Setup file logging
+    # USER_LOG_FILE dir must exist
+    USER_LOG_FILE.parent.mkdir(parents=True, exist_ok=True)
+    file_handler = logging.FileHandler(USER_LOG_FILE)
+    file_formatter = logging.Formatter(
+        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
+    file_handler.setLevel(logging.DEBUG)
+    file_handler.setFormatter(file_formatter)
+    logger.addHandler(file_handler)
+
+    # Needs to be DEBUG otherwise will filter before sending to handlers
+    logger.setLevel(logging.DEBUG)
 
     global FORCE_COLOR
     if force_color: